+++ client/app/reader/constants.js
@@ -7,7 +7,6 @@ export const TOGGLE_DOCUMENT_CATEGORY_FAIL = 'TOGGLE_DOCUMENT_CATEGORY_FAIL';
 export const RECEIVE_DOCUMENTS = 'RECEIVE_DOCUMENTS';
 export const RECEIVE_MANIFESTS = 'RECEIVE_MANIFESTS';
 export const RECEIVE_ANNOTATIONS = 'RECEIVE_ANNOTATIONS';
-export const RECEIVE_ASSIGNMENTS = 'RECEIVE_ASSIGNMENTS';
 export const RECEIVE_APPEAL_DETAILS = 'RECEIVE_APPEAL_DETAILS';
 export const RECEIVE_APPEAL_DETAILS_FAILURE = 'RECEIVE_APPEAL_DETAILS_FAILURE';
 export const REQUEST_INITIAL_DATA_FAILURE = 'INITIAL_DATA_LOADING_FAIL';
